Extract ConceptDiagram into a component and allow css to be injected
Requirements for width, alignment are diffent on pipelines page and elastic agents page
