+++ server/src/main/webapp/WEB-INF/rails/webpack/views/components/concept_diagram/index.scss
