+++ server/src/main/webapp/WEB-INF/rails/webpack/views/components/concept_diagram/index.scss.d.ts
@@ -1,0 +1,2 @@
+// This file is automatically generated. Please do not change this file!
+export const conceptDiagram: string;
